The death of AutoDrive...
RIP AutoDrive 2/15/17
It was replaced with PID
He is survived by his brother, AutoTurn, and son, AutoTest. His wife,
AutoLaunch, is still mourning the loss.
His cousins, AutoGear and AutoGearFuelHighCross, also died tonight. We
will all remember this travesty, in our hearts.
All of his friends are doing their best to move on and learn to live
with PID.
1 push = 1 prayer